# Conflicts:
#	backend/src/main/java/com/ideahub/dao/IdeaDAO.java
#	backend/src/main/java/com/ideahub/resources/IdeaResource.java

+++ backend/src/test/java/com/ideahub/resources/IdeaCollaboratorResourceTest.java
@@ -92,7 +92,7 @@ public class IdeaCollaboratorResourceTest {
 
         IdeaCollaboratorId collabId = IdeaCollaboratorId.builder()
                 .userId((int) userB.getId())
-                .ideaId((int) idea.getId())
+                .ideaId(idea.getId())
                 .build();
         final IdeaInvitation invitation = IdeaInvitation.builder()
                 .acceptedState(false)
